gift_app/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
allure-pytest>=2.13.0
httpx>=0.24.0
requests>=2.25.0
ruff>=0.1.0
black>=23.0.0
isort>=5.12.0
pre-commit>=3.0.0
mypy>=1.0.0
types-requests
